+++ spec/data/node_spec.rb
@@ -372,7 +372,7 @@ describe Razor::Data::Node do
     end
   end
 
-  context "ipmi" do
+  context "IPMI" do
     context "validation" do
       it "should work with only hostname set" do
         # expect this to raise no errors
@@ -448,7 +448,7 @@ describe Razor::Data::Node do
     end
 
     describe "last_known_power_state" do
-      [true, false, nil].each do |value|
+      ['on', 'off', nil].each do |value|
         it "should update the timestamp when the power state is set to #{value.inspect}" do
           Timecop.freeze do
             # Make sure that we are not mis-detecting a previously set value
@@ -466,12 +466,15 @@ describe Razor::Data::Node do
     end
 
     describe "update_power_state!" do
+      include TorqueBox::Injectors
+
       before :each do
         node.update(:ipmi_hostname => Faker::Internet.domain_name).save
       end
 
       let :prehistory do Time.at(-446061360) end
       let :thefuture  do Time.at(1445480880) end
+      let :queue      do fetch('/queues/razor/sequel-instance-messages') end
 
 
       # @todo danielp 2013-12-05: this stubs the IPMI on? question, which is
@@ -479,9 +482,9 @@ describe Razor::Data::Node do
       # run, or the power state query mechanism, both of which are about as
       # tied to implementation, so I don't feel too bad about this...
 
-      { 'on' => true, 'off' => false }.each do |name, value|
-        it "should update the power state if the machine is #{name}" do
-          Razor::IPMI.stub(:on?).and_return(value)
+      %w{on off}.each do |value|
+        it "should update the power state if the machine is #{value}" do
+          Razor::IPMI.stub(:on?).and_return(value == 'on')
 
           Timecop.freeze(prehistory) do
             node.last_known_power_state.should be_nil
@@ -489,7 +492,7 @@ describe Razor::Data::Node do
 
             node.update_power_state!
 
-            node.last_known_power_state.should(value ? be_true : be_false)
+            node.last_known_power_state.should == value
             node.last_power_state_update_at.should == prehistory
           end
         end
@@ -504,12 +507,12 @@ describe Razor::Data::Node do
 
           node.update_power_state!
 
-          node.last_known_power_state.should be_true
+          node.last_known_power_state.should == 'on'
           node.last_power_state_update_at.should == prehistory
 
           Timecop.freeze(thefuture) do
             node.update_power_state!
-            node.last_known_power_state.should be_true
+            node.last_known_power_state.should == 'on'
             node.last_power_state_update_at.should == thefuture
           end
         end
@@ -524,7 +527,7 @@ describe Razor::Data::Node do
 
           node.update_power_state!
 
-          node.last_known_power_state.should be_true
+          node.last_known_power_state.should == 'on'
           node.last_power_state_update_at.should == prehistory
 
           Timecop.freeze(thefuture) do
@@ -550,7 +553,7 @@ describe Razor::Data::Node do
 
           node.update_power_state!
 
-          node.last_known_power_state.should be_true
+          node.last_known_power_state.should == 'on'
           node.last_power_state_update_at.should == prehistory
 
           Timecop.freeze(thefuture) do
@@ -561,10 +564,87 @@ describe Razor::Data::Node do
               node.update_power_state!
             }.to raise_error exception.class, exception.message
 
-            node.last_known_power_state.should be_true
+            node.last_known_power_state.should == 'on'
             node.last_power_state_update_at.should == prehistory
           end
         end
+      end
+
+      it "should queue nothing else if desired power state is null" do
+        Razor::IPMI.stub(:on?).and_return(true, false)
+
+        # First, it is on...
+        node.update_power_state!
+        queue.should be_empty
+
+        # ...then it is off...
+        node.update_power_state!
+        queue.should be_empty
+      end
+
+      it "should queue nothing if desired is off and actual is off" do
+        Razor::IPMI.stub(:on?).and_return(false)
+        node.set(:desired_power_state => 'off').save
+        node.update_power_state!
+        queue.should be_empty
+      end
+
+      it "should queue nothing if desired is on and actual is on" do
+        Razor::IPMI.stub(:on?).and_return(true)
+        node.set(:desired_power_state => 'on').save
+        node.update_power_state!
+        queue.should be_empty
+      end
+
+      it "should queue turning off if desired is off and actual is on" do
+        Razor::IPMI.stub(:on?).and_return(true)
+        node.set(:desired_power_state => 'off').save
+
+        expect {
+          node.update_power_state!
+        }.to have_published(
+          'class'     => node.class.name,
+          'instance'  => node.pk_hash,
+          'message'   => 'off',
+          'arguments' => []
+        ).on(queue)
+      end
+
+      it "should queue turning on if desired is on and actual is off" do
+        Razor::IPMI.stub(:on?).and_return(false)
+        node.set(:desired_power_state => 'on').save
+
+        expect {
+          node.update_power_state!
+        }.to have_published(
+          'class'     => node.class.name,
+          'instance'  => node.pk_hash,
+          'message'   => 'on',
+          'arguments' => []
+        ).on(queue)
+      end
+
+      it "should queue turning on twice if actual remains off when checked again" do
+        Razor::IPMI.stub(:on?).and_return(false)
+        node.set(:desired_power_state => 'on').save
+
+        expect {
+          node.update_power_state!
+        }.to have_published(
+          'class'     => node.class.name,
+          'instance'  => node.pk_hash,
+          'message'   => 'on',
+          'arguments' => []
+        ).on(queue)
+
+        expect {
+          node.update_power_state!
+        }.to have_published(
+          'class'     => node.class.name,
+          'instance'  => node.pk_hash,
+          'message'   => 'on',
+          'arguments' => []
+        ).on(queue)
       end
     end
   end